# docs/contracts/python-openapi.json
{
  "components": {
    "schemas": {
      "AccountBalanceIn": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "integer"
          },
          "balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "account_id"
        ],
        "title": "AccountBalanceIn",
        "type": "object"
      },
      "AccountBalanceOut": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "integer"
          },
          "balance": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "account_id",
          "balance",
          "notes"
        ],
        "title": "AccountBalanceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NetWorthSeriesPoint": {
        "properties": {
          "by_account": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "By Account",
            "type": "object"
          },
          "by_category": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "By Category",
            "type": "object"
          },
          "snapshot_date": {
            "format": "date",
            "title": "Snapshot Date",
            "type": "string"
          },
          "tax_advantaged": {
            "title": "Tax Advantaged",
            "type": "number"
          },
          "taxable": {
            "title": "Taxable",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "snapshot_date",
          "total",
          "by_category",
          "by_account",
          "taxable",
          "tax_advantaged"
        ],
        "title": "NetWorthSeriesPoint",
        "type": "object"
      },
      "NetWorthSnapshotIn": {
        "properties": {
          "balances": {
            "items": {
              "$ref": "#/components/schemas/AccountBalanceIn"
            },
            "title": "Balances",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "snapshot_date": {
            "format": "date",
            "title": "Snapshot Date",
            "type": "string"
          }
        },
        "required": [
          "snapshot_date"
        ],
        "title": "NetWorthSnapshotIn",
        "type": "object"
      },
      "NetWorthSnapshotOut": {
        "properties": {
          "balances": {
            "items": {
              "$ref": "#/components/schemas/AccountBalanceOut"
            },
            "title": "Balances",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "snapshot_date": {
            "format": "date",
            "title": "Snapshot Date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "snapshot_date",
          "notes",
          "balances"
        ],
        "title": "NetWorthSnapshotOut",
        "type": "object"
      },
      "NetWorthSnapshotUpdate": {
        "properties": {
          "balances": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AccountBalanceIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balances"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "snapshot_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Date"
          }
        },
        "title": "NetWorthSnapshotUpdate",
        "type": "object"
      },
      "NetWorthWorkbookImportRequest": {
        "properties": {
          "account_map": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Account Map",
            "type": "object"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "NetWorthWorkbookImportRequest",
        "type": "object"
      },
      "NetWorthWorkbookImportResult": {
        "properties": {
          "imported": {
            "title": "Imported",
            "type": "integer"
          },
          "missing_accounts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Missing Accounts",
            "type": "array"
          },
          "skipped_existing": {
            "title": "Skipped Existing",
            "type": "integer"
          }
        },
        "required": [
          "imported",
          "skipped_existing"
        ],
        "title": "NetWorthWorkbookImportResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DeskBooks",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/snapshots": {
      "get": {
        "operationId": "list_snapshots_api_snapshots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetWorthSnapshotOut"
                  },
                  "title": "Response List Snapshots Api Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Snapshots",
        "tags": [
          "snapshots"
        ]
      },
      "post": {
        "operationId": "create_snapshot_api_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetWorthSnapshotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthSnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Snapshot",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/snapshots/import-workbook": {
      "post": {
        "operationId": "import_workbook_api_snapshots_import_workbook_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetWorthWorkbookImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthWorkbookImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Workbook",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/snapshots/series": {
      "get": {
        "operationId": "snapshot_series_api_snapshots_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetWorthSeriesPoint"
                  },
                  "title": "Response Snapshot Series Api Snapshots Series Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Snapshot Series",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/snapshots/{snap_id}": {
      "delete": {
        "operationId": "delete_snapshot_api_snapshots__snap_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "snap_id",
            "required": true,
            "schema": {
              "title": "Snap Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Snapshot",
        "tags": [
          "snapshots"
        ]
      },
      "patch": {
        "operationId": "update_snapshot_api_snapshots__snap_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "snap_id",
            "required": true,
            "schema": {
              "title": "Snap Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetWorthSnapshotUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthSnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Snapshot",
        "tags": [
          "snapshots"
        ]
      }
    },
...
  }
}